requirements.txt
torch
torchvision
numpy
matplotlib
scikit-image
pillow